refactor(comms/specs): collapse worker specs into single module file
