--- comms/src/specs/worker/algorithm.rs
@@ -1,12 +1,0 @@
-use std::net::SocketAddr;
-
-use serde::{Deserialize, Serialize};
-
-/// Distributed training algorithm selection.
-///
-/// Only `parameter_server` is currently implemented.
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum AlgorithmSpec {
-    ParameterServer { server_ip: SocketAddr },
-}
