--- comms/src/specs/worker/model.rs
@@ -1,36 +1,0 @@
-use std::num::NonZeroUsize;
-
-use serde::{Deserialize, Serialize};
-
-/// Model selection and configuration.
-///
-/// The initial variants exist to keep the system testable while the model
-/// runtime is expanded incrementally.
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ModelSpec {
-    /// No-op model used for protocol tests.
-    Noop,
-    /// Mock model used for end-to-end tests (gradient = 2 * weights).
-    Mock,
-    /// Feed-forward network (placeholder for the next refactor steps).
-    FeedForward { layers: Vec<LayerSpec> },
-}
-
-/// Feed-forward layer specification.
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum LayerSpec {
-    Dense {
-        input: NonZeroUsize,
-        output: NonZeroUsize,
-        act_fn: Option<ActFnSpec>,
-    },
-}
-
-/// Activation function specification.
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ActFnSpec {
-    Sigmoid { amp: f32 },
-}
